tambah middleware

+++ routes/web.php
@@ -50,52 +50,59 @@ Route::name('practice.')->group(function () {
 // Route::get('produk/create', [ProductController::class, 'create']);
 
 //dashboard
-Route::get('dashboard', [DashAdminController::class, 'index']);
-Route::get('dashboard', [ProductController::class, 'HitungData']);
-Route::get('dashboard', [StockController::class, 'stokForAdminDash']);
+Route::get('dashboard', [DashAdminController::class, 'index'])->middleware('userAkses:admin');
+Route::get('dashboard', [ProductController::class, 'HitungData'])->middleware('userAkses:admin');
+Route::get('dashboard', [StockController::class, 'stokForAdminDash'])->middleware('userAkses:admin');
 
-
-
-Route::get('dashboard/manager', [DashAdminController::class, 'managerIndex']);
-Route::get('dashboard/staf', [DashAdminController::class, 'stafIndex']);  
-Route::get('dashboard/staf', [StockController::class, 'perluDiPeriksaStaf']); 
+Route::get('dashboard/manager', [DashAdminController::class, 'managerIndex'])->middleware('userAkses:manager_gudang');
+Route::get('dashboard/staf', [DashAdminController::class, 'stafIndex'])->middleware('userAkses:staf_gudang');  
+Route::get('dashboard/staf', [StockController::class, 'perluDiPeriksaStaf'])->middleware('userAkses:staf_gudang'); 
 
 //produk
 Route::resource('produk', ProductController::class);
-
 //kategori
-Route::resource('kategori', KategoriController::class);
-
+Route::resource('kategori', KategoriController::class)->middleware('userAkses:admin');
 //suppliers
-Route::resource('supplier',SupplierController::class);
+Route::resource('supplier',SupplierController::class)->middleware('userAkses:admin');
+Route::get('users', [UserController::class, 'index'])->name('users')->middleware('userAkses:admin');
+Route::delete('user-delete/{id}', [UserController::class, 'destroy'])->name('users.destroy')->middleware('userAkses:admin');
+Route::get('user-edit/{id}', [UserController::class, 'edit'])->middleware('userAkses:admin');
+Route::post('user-edit/{id}', [UserController::class, 'update'])->middleware('userAkses:admin');
 
 //manager
-Route::resource('produkManager', ProductManagerDashboard::class);
-Route::get('supplierManager', [SupplierController::class, 'supplierForManager']);
-Route::get('dashboard/manager', [StockController::class, 'stokForManager']);
+Route::resource('produkManager', ProductManagerDashboard::class)->middleware('userAkses:manager_gudang');
+Route::get('supplierManager', [SupplierController::class, 'supplierForManager'])->middleware('userAkses:manager_gudang');
+Route::get('dashboard/manager', [StockController::class, 'stokForManager'])->middleware('userAkses:manager_gudang');
 
 //stocktransaction
-Route::get('/stok', [StockController::class, 'index'])->name('stok');
-Route::get('/stokAdmin', [StockController::class, 'stokForAdmin'])->name('stokAdmin');
-Route::get('stok-create', [StockController::class, 'create']);
-Route::post('stok-create', [StockController::class, 'store']);
-Route::get('stok-edit/{id}', [StockController::class, 'edit']);
-Route::post('stok-edit/{id}', [StockController::class, 'update']);
-Route::delete('stok-delete/{id}', [StockController::class, 'destroy'])->name('stok.destroy');
-Route::get('dashboard/{type}', [StockController::class, 'getStockByType']);
-Route::get('periksa', [StockController::class, 'getTransByStatus']);
+Route::get('/stok', [StockController::class, 'index'])->name('stok')->middleware('userAkses:manager_gudang');
+Route::get('/stokAdmin', [StockController::class, 'stokForAdmin'])->name('stokAdmin')->middleware('userAkses:admin');
+Route::get('stok-create', [StockController::class, 'create'])->middleware('userAkses:manager_gudang');
+Route::post('stok-create', [StockController::class, 'store'])->middleware('userAkses:manager_gudang');
+Route::get('stok-edit/{id}', [StockController::class, 'edit'])->middleware('userAkses:manager_gudang');
+Route::post('stok-edit/{id}', [StockController::class, 'update'])->middleware('userAkses:manager_gudang');
+Route::delete('stok-delete/{id}', [StockController::class, 'destroy'])->name('stok.destroy')->middleware('userAkses:manager_gudang');
+Route::get('dashboard/{type}', [StockController::class, 'getStockByType'])->middleware('userAkses:manager_gudang');
+
+Route::get('periksa', [StockController::class, 'getTransByStatus'])->middleware('userAkses:staf_gudang');
 
 //login-reg
-Route::get('sesi', [SesiController::class, 'index']);
-Route::post('sesi', [SesiController::class, 'login']);
-Route::get('/logout', [SesiController::class, 'logout']);
-Route::get('sesi/register', [SesiController::class, 'reg']);
-Route::post('sesi/register', [SesiController::class, 'register']);
+Route::middleware(['guest'])->group(function(){
+    Route::get('sesi', [SesiController::class, 'index'])->name('login');
+    Route::post('sesi', [SesiController::class, 'login']);
+    Route::get('sesi/register', [SesiController::class, 'reg']);
+    Route::post('sesi/register', [SesiController::class, 'register']); 
+});
+Route::get('/home', function(){
+    return redirect('dashboard/staf');
+});
+
+
+Route::middleware(['auth'])->group(function(){
+    Route::get('/logout', [SesiController::class, 'logout']);
+});
 
 //
-Route::get('users', [UserController::class, 'index'])->name('users');
-Route::delete('user-delete/{id}', [UserController::class, 'destroy'])->name('users.destroy');
-Route::get('user-edit/{id}', [UserController::class, 'edit']);
-Route::post('user-edit/{id}', [UserController::class, 'update']);
+
